+++ app/jobs/mentions/create_all_job.rb
@@ -3,8 +3,10 @@ module Mentions
     queue_as :mentions_create_all
 
     def perform(notifiable_id, notifiable_type)
+      return if ["Comment"].none?(notifiable_type)
+
       notifiable = notifiable_type.constantize.find_by(id: notifiable_id)
-      MentionsCreateAllService.call(notifiable) if notifiable
+      Mentions::CreateAll.call(notifiable) if notifiable
     end
   end
 end
